Reorganize texture loader code
Clean up some code
Isolate Basis Universal loader into a separate module
