+++ src/resource/texture_loader/meson.build
@@ -1,0 +1,5 @@
+
+resource_texload_src = files(
+    'texture_loader.c',
+    'texture_loader_basisu.c',
+)
